+++ include/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+# 普通题库
+add_subdirectory(dataset)
+
+# leetcode 题库
+add_subdirectory(leetcode)
