+++ include/dataset/3_package.h
@@ -1,7 +1,12 @@
-#include <iostream>
-#include <vector>
-using namespace std;
-class Problem_3 {
+#include "problem.h"
+/*
+问题概述: 
+问题详细描述: 
+解法:
+状态: pass
+*/
+
+class Problem_3 : public Problem {
 public:
     int FindMaxValueInner(int start, int currentValue, int remainSize, char *isVisited)
     {
@@ -16,14 +21,15 @@ public:
         }
         return ret;
     }
+
     int FindMaxValue(int volumn)
     {
         int N = value.size();
         int M = weight.size();
         char *isVisited = new char[N];
         memset(isVisited, 0, sizeof(char) * N);
-        int ret = FindMaxValueInner(0, volumn, isVisited);
-        delete []isVisited;
+        int ret = FindMaxValueInner(0, 0, volumn, isVisited);
+        delete [] isVisited;
         return ret;
     }
