+++ include/dataset/problem.h
@@ -1,0 +1,15 @@
+#include <iostream>
+#include <vector>
+#include <string>
+using namespace std;
+/*
+问题概述: 
+问题详细描述: 
+解法:
+状态: pass
+*/
+class Problem {
+public:
+    Problem() {};
+   virtual ~Problem() {};
+};
