--- include/leetcode/0_Template.h
@@ -1,6 +1,0 @@
-#include <iostream>
-using namespace std;
-class Solution {
-public:
-    
-};
