+++ include/leetcode/Solution.h
@@ -1,0 +1,6 @@
+#include <iostream>
+using namespace std;
+class Solution {
+public:
+    
+};
