+++ pub/type.h
@@ -15,10 +15,13 @@ extern "C" {
 #endif
 
 typedef char int8;
-typedef long long int int64;
+typedef unsigned char uint8;
 
 typedef unsigned int uint;
-typedef unsigned char uint8;
+typedef unsigned int uint32;
+typedef int int32;
+
+typedef long long int int64;
 typedef unsigned long long int uint64;
 
 #ifdef __cplusplus
