+++ test/dt/ut_dataset/dataset_test.cc
@@ -1,0 +1,2 @@
+#include "dataset_pub.h"
+
