--- test/llt/ut/ut_testbook/testbook.cc
@@ -1,2 +1,0 @@
-#include "testbook_pub.h"
-
